+++ projects/js-packages/charts/src/components/pie-chart/pie-chart.tsx
@@ -1,17 +1,18 @@
 import { Group } from '@visx/group';
 import { Pie } from '@visx/shape';
 import clsx from 'clsx';
-import { SVGProps } from 'react';
+import { SVGProps, type MouseEvent } from 'react';
 import useChartMouseHandler from '../../hooks/use-chart-mouse-handler';
 import { useChartTheme, defaultTheme } from '../../providers/theme';
 import { Legend } from '../legend';
+import { withResponsive } from '../shared/with-responsive';
 import { BaseTooltip } from '../tooltip';
 import styles from './pie-chart.module.scss';
-import type { BaseChartProps, DataPoint } from '../shared/types';
+import type { BaseChartProps, DataPointPercentage } from '../../types';
 
 // TODO: add animation
 
-interface PieChartProps extends BaseChartProps< DataPoint[] > {
+interface PieChartProps extends BaseChartProps< DataPointPercentage[] > {
 	/**
 	 * Inner radius in pixels. If > 0, creates a donut chart. Defaults to 0.
 	 */
@@ -26,8 +27,8 @@ interface PieChartProps extends BaseChartProps< DataPoint[] > {
  */
 const PieChart = ( {
 	data,
-	width,
-	height,
+	width = 500, //TODO: replace when making the components responsive
+	height = 500, //TODO: replace when making the components responsive
 	withTooltips = false,
 	innerRadius = 0,
 	className,
@@ -45,10 +46,17 @@ const PieChart = ( {
 	const centerX = width / 2;
 	const centerY = height / 2;
 
+	// Map the data to include index for color assignment
+	const dataWithIndex = data.map( ( d, index ) => ( {
+		...d,
+		index,
+	} ) );
+
 	const accessors = {
-		value: d => d.value,
+		value: ( d: DataPointPercentage ) => d.value,
 		// Use the color property from the data object as a last resort. The theme provides colours by default.
-		fill: d => d.color || providerTheme.colors[ d.index ],
+		fill: ( d: DataPointPercentage & { index: number } ) =>
+			d?.color || providerTheme.colors[ d.index ],
 	};
 
 	// Create legend items from data
@@ -62,8 +70,8 @@ const PieChart = ( {
 		<div className={ clsx( 'pie-chart', styles[ 'pie-chart' ], className ) }>
 			<svg width={ width } height={ height }>
 				<Group top={ centerY } left={ centerX }>
-					<Pie
-						data={ data }
+					<Pie< DataPointPercentage & { index: number } >
+						data={ dataWithIndex }
 						pieValue={ accessors.value }
 						outerRadius={ radius - 20 } // Leave space for labels/tooltips
 						innerRadius={ innerRadius }
@@ -72,11 +80,12 @@ const PieChart = ( {
 							return pie.arcs.map( ( arc, index ) => {
 								const [ centroidX, centroidY ] = pie.path.centroid( arc );
 								const hasSpaceForLabel = arc.endAngle - arc.startAngle >= 0.25;
-								const handleMouseMove = event => onMouseMove( event, arc.data );
+								const handleMouseMove = ( event: MouseEvent< SVGElement > ) =>
+									onMouseMove( event, arc.data );
 
 								const pathProps: SVGProps< SVGPathElement > = {
 									d: pie.path( arc ) || '',
-									fill: accessors.fill( arc ),
+									fill: accessors.fill( arc.data ),
 								};
 
 								if ( withTooltips ) {
@@ -121,8 +130,8 @@ const PieChart = ( {
 			{ withTooltips && tooltipOpen && tooltipData && (
 				<BaseTooltip
 					data={ tooltipData }
-					top={ tooltipTop }
-					left={ tooltipLeft }
+					top={ tooltipTop || 0 }
+					left={ tooltipLeft || 0 }
 					style={ {
 						transform: 'translate(-50%, -100%)',
 					} }
@@ -132,4 +141,5 @@ const PieChart = ( {
 	);
 };
 
-export default PieChart;
+PieChart.displayName = 'PieChart';
+export default withResponsive< PieChartProps >( PieChart );