+++ projects/packages/my-jetpack/_inc/components/product-cards-section/all.ts
@@ -5,9 +5,12 @@ import BoostCard from './boost-card';
 import CompleteCard from './complete-card';
 import CrmCard from './crm-card';
 import GrowthCard from './growth-card';
+import NewsletterCard from './newsletter-card';
 import ProtectCard from './protect-card';
+import RelatedPostsCard from './related-posts-card';
 import SearchCard from './search-card';
 import SecurityCard from './security-card';
+import SiteAcceleratorCard from './site-accelerator-card';
 import SocialCard from './social-card';
 import StatsCard from './stats-card';
 import VideopressCard from './videopress-card';
@@ -33,4 +36,8 @@ export const JetpackModuleToProductCard: {
 	extras: null,
 	scan: null,
 	creator: null,
+	// Features:
+	newsletter: NewsletterCard,
+	'related-posts': RelatedPostsCard,
+	'site-accelerator': SiteAcceleratorCard,
 };