+++ projects/plugins/social/tests/e2e/specs/admin-page.test.js
@@ -1,7 +1,6 @@
 import { prerequisitesBuilder } from 'jetpack-e2e-commons/env/prerequisites.js';
 import { expect, test } from 'jetpack-e2e-commons/fixtures/base-test.js';
 import logger from 'jetpack-e2e-commons/logger.js';
-import { JetpackSocialPage } from '../pages/index.js';
 
 test.beforeEach( async ( { page } ) => {
 	await prerequisitesBuilder( page )
@@ -11,9 +10,10 @@ test.beforeEach( async ( { page } ) => {
 		.build();
 } );
 
-test( 'Jetpack Social admin page', async ( { page } ) => {
+test( 'Jetpack Social admin page', async ( { page, admin } ) => {
 	logger.action( 'Visit the Jetpack Social admin page' );
-	await JetpackSocialPage.visit( page );
+
+	await admin.visitAdminPage( 'admin.php', 'page=jetpack-social' );
 
 	logger.action( 'Checking for heading "Jetpack Social"' );
 	await expect( page.getByRole( 'heading', { name: 'Jetpack Social' } ) ).toBeVisible();