+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.2.2] - 2025-01-06
+### Added
+- Added tests to increase code coverage. [#39963]
+
+### Changed
+- Updated package dependencies. [#40831]
+
 ## [6.2.1] - 2024-12-16
 ### Changed
 - Updated package dependencies. [#40564]
@@ -1271,6 +1278,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[6.2.2]: https://github.com/Automattic/jetpack-connection/compare/v6.2.1...v6.2.2
 [6.2.1]: https://github.com/Automattic/jetpack-connection/compare/v6.2.0...v6.2.1
 [6.2.0]: https://github.com/Automattic/jetpack-connection/compare/v6.1.1...v6.2.0
 [6.1.1]: https://github.com/Automattic/jetpack-connection/compare/v6.1.0...v6.1.1